attach

+++ Client/src/main/java/gov/iti/jets/dto/AttachementDTO.java
@@ -5,7 +5,7 @@ public class AttachementDTO {
     private int attachmentID;
     private String attachmentTitle;
     private String attachmentType;
-    private int attachmentSize;
+    private long attachmentSize;
 
 
     /**
@@ -53,14 +53,14 @@ public class AttachementDTO {
     /**
      * @return int return the attachmentSize
      */
-    public int getAttachmentSize() {
+    public long getAttachmentSize() {
         return attachmentSize;
     }
 
     /**
      * @param attachmentSize the attachmentSize to set
      */
-    public void setAttachmentSize(int attachmentSize) {
+    public void setAttachmentSize(long attachmentSize) {
         this.attachmentSize = attachmentSize;
     }
